pyproject.toml
[project]
name = 'parseman'
description = 'A parser for typed parameters (command line arguments, yaml files, json files, etc)'
keywords = ['parser', 'cli', 'yaml', 'json', 'schema']
classifiers = [
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3',
    'Natural Language :: English',
    'Topic :: Scientific/Engineering',
]
requires-python = '>=3.6'
dynamic = ['version', 'readme']
dependencies = [
    'pyyaml',
    'jsonschema'
]

[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'

[tool.setuptools.dynamic]
readme = {file = ['readme.md'], content-type = 'text/markdown'}
version = {attr = 'parseman.__version__'}
